// packages/services/shopify/src/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Collection = HasMetafields & Node & OnlineStorePublishable & Trackable & {
  readonly __typename: 'Collection';
  /** Stripped description of the collection, single line with HTML tags removed. */
  readonly description: Scalars['String']['output'];
  /** The description of the collection, complete with HTML formatting. */
  readonly descriptionHtml: Scalars['HTML']['output'];
  /**
   * A human-friendly unique string for the collection automatically generated from its title.
   * Limit of 255 characters.
   *
   */
  readonly handle: Scalars['String']['output'];
  /** A globally-unique ID. */
  readonly id: Scalars['ID']['output'];
  /** Image associated with the collection. */
  readonly image: Maybe<Image>;
  /** A [custom field](https://shopify.dev/docs/apps/build/custom-data), including its `namespace` and `key`, that's associated with a Shopify resource for the purposes of adding and storing additional information. */
  readonly metafield: Maybe<Metafield>;
  /** A list of [custom fields](/docs/apps/build/custom-data) that a merchant associates with a Shopify resource. */
  readonly metafields: ReadonlyArray<Maybe<Metafield>>;
  /** The URL used for viewing the resource on the shop's Online Store. Returns `null` if the resource is currently not published to the Online Store sales channel. */
  readonly onlineStoreUrl: Maybe<Scalars['URL']['output']>;
  /** List of products in the collection. */
  readonly products: ProductConnection;
  /** The collection's SEO information. */
  readonly seo: Seo;
  /** The collection’s name. Limit of 255 characters. */
  readonly title: Scalars['String']['output'];
  /** URL parameters to be added to a page URL to track the origin of on-site search traffic for [analytics reporting](https://help.shopify.com/manual/reports-and-analytics/shopify-reports/report-types/default-reports/behaviour-reports). Returns a result when accessed through the [search](https://shopify.dev/docs/api/storefront/current/queries/search) or [predictiveSearch](https://shopify.dev/docs/api/storefront/current/queries/predictiveSearch) queries, otherwise returns null. */
  readonly trackingParameters: Maybe<Scalars['String']['output']>;
  /** The date and time when the collection was last modified. */
  readonly updatedAt: Scalars['DateTime']['output'];
};
// ...
export type Image = {
  readonly __typename: 'Image';
  /** A word or phrase to share the nature or contents of an image. */
  readonly altText: Maybe<Scalars['String']['output']>;
  /** The original height of the image in pixels. Returns `null` if the image isn't hosted by Shopify. */
  readonly height: Maybe<Scalars['Int']['output']>;
  /** A unique ID for the image. */
  readonly id: Maybe<Scalars['ID']['output']>;
  /**
   * The location of the original image as a URL.
   *
   * If there are any existing transformations in the original source URL, they will remain and not be stripped.
   *
   * @deprecated Use `url` instead.
   */
  readonly originalSrc: Scalars['URL']['output'];
  /**
   * The location of the image as a URL.
   * @deprecated Use `url` instead.
   */
  readonly src: Scalars['URL']['output'];
  /**
   * The ThumbHash of the image.
   *
   * Useful to display placeholder images while the original image is loading.
   *
   * See https://evanw.github.io/thumbhash/ for details on how to use it.
   *
   */
  readonly thumbhash: Maybe<Scalars['String']['output']>;
  /**
   * The location of the transformed image as a URL.
   *
   * All transformation arguments are considered "best-effort". If they can be applied to an image, they will be.
   * Otherwise any transformations which an image type doesn't support will be ignored.
   *
   * @deprecated Use `url(transform:)` instead
   */
  readonly transformedSrc: Scalars['URL']['output'];
  /**
   * The location of the image as a URL.
   *
   * If no transform options are specified, then the original image will be preserved including any pre-applied transforms.
   *
   * All transformation options are considered "best-effort". Any transformation that the original image type doesn't support will be ignored.
   *
   * If you need multiple variations of the same image, then you can use [GraphQL aliases](https://graphql.org/learn/queries/#aliases).
   *
   */
  readonly url: Scalars['URL']['output'];
  /** The original width of the image in pixels. Returns `null` if the image isn't hosted by Shopify. */
  readonly width: Maybe<Scalars['Int']['output']>;
};
// ...
export type MoneyV2 = {
  readonly __typename: 'MoneyV2';
  /** Decimal money amount. */
  readonly amount: Scalars['Decimal']['output'];
  /** Currency of the money. */
  readonly currencyCode: CurrencyCode;
};
// ...
export type Node = {
  /** A globally-unique ID. */
  readonly id: Scalars['ID']['output'];
};
// ...
export type PageInfo = {
  readonly __typename: 'PageInfo';
  /** The cursor corresponding to the last node in edges. */
  readonly endCursor: Maybe<Scalars['String']['output']>;
  /** Whether there are more pages to fetch following the current page. */
  readonly hasNextPage: Scalars['Boolean']['output'];
  /** Whether there are any pages prior to the current page. */
  readonly hasPreviousPage: Scalars['Boolean']['output'];
  /** The cursor corresponding to the first node in edges. */
  readonly startCursor: Maybe<Scalars['String']['output']>;
};
// ...
export type ProductConnection = {
  readonly __typename: 'ProductConnection';
  /** A list of edges. */
  readonly edges: ReadonlyArray<ProductEdge>;
  /** A list of available filters. */
  readonly filters: ReadonlyArray<Filter>;
  /** A list of the nodes contained in ProductEdge. */
  readonly nodes: ReadonlyArray<Product>;
  /** Information to aid in pagination. */
  readonly pageInfo: PageInfo;
};

/**
 * An auto-generated type which holds one Product and a cursor during pagination.
 *
 */
export type ProductEdge = {
  readonly __typename: 'ProductEdge';
  /** A cursor for use in pagination. */
  readonly cursor: Scalars['String']['output'];
  /** The item at the end of ProductEdge. */
  readonly node: Product;
};
// ...
export type ProductVariant = HasMetafields & Node & {
  readonly __typename: 'ProductVariant';
  /** Indicates if the product variant is available for sale. */
  readonly availableForSale: Scalars['Boolean']['output'];
  /** The barcode (for example, ISBN, UPC, or GTIN) associated with the variant. */
  readonly barcode: Maybe<Scalars['String']['output']>;
  /** The compare at price of the variant. This can be used to mark a variant as on sale, when `compareAtPrice` is higher than `price`. */
  readonly compareAtPrice: Maybe<MoneyV2>;
  /**
   * The compare at price of the variant. This can be used to mark a variant as on sale, when `compareAtPriceV2` is higher than `priceV2`.
   * @deprecated Use `compareAtPrice` instead.
   */
  readonly compareAtPriceV2: Maybe<MoneyV2>;
  /**
   * List of bundles components included in the variant considering only fixed bundles.
   *
   */
  readonly components: ProductVariantComponentConnection;
  /** Whether a product is out of stock but still available for purchase (used for backorders). */
  readonly currentlyNotInStock: Scalars['Boolean']['output'];
  /**
   * List of bundles that include this variant considering only fixed bundles.
   *
   */
  readonly groupedBy: ProductVariantConnection;
  /** A globally-unique ID. */
  readonly id: Scalars['ID']['output'];
  /** Image associated with the product variant. This field falls back to the product image if no image is available. */
  readonly image: Maybe<Image>;
  /** A [custom field](https://shopify.dev/docs/apps/build/custom-data), including its `namespace` and `key`, that's associated with a Shopify resource for the purposes of adding and storing additional information. */
  readonly metafield: Maybe<Metafield>;
  /** A list of [custom fields](/docs/apps/build/custom-data) that a merchant associates with a Shopify resource. */
  readonly metafields: ReadonlyArray<Maybe<Metafield>>;
  /** The product variant’s price. */
  readonly price: MoneyV2;
  /**
   * The product variant’s price.
   * @deprecated Use `price` instead.
   */
  readonly priceV2: MoneyV2;
  /** The product object that the product variant belongs to. */
  readonly product: Product;
  /** The total sellable quantity of the variant for online sales channels. */
  readonly quantityAvailable: Maybe<Scalars['Int']['output']>;
  /** A list of quantity breaks for the product variant. */
  readonly quantityPriceBreaks: QuantityPriceBreakConnection;
  /** The quantity rule for the product variant in a given context. */
  readonly quantityRule: QuantityRule;
  /**
   * Whether a product variant requires components. The default value is `false`.
   * If `true`, then the product variant can only be purchased as a parent bundle with components.
   *
   */
  readonly requiresComponents: Scalars['Boolean']['output'];
  /** Whether a customer needs to provide a shipping address when placing an order for the product variant. */
  readonly requiresShipping: Scalars['Boolean']['output'];
  /** List of product options applied to the variant. */
  readonly selectedOptions: ReadonlyArray<SelectedOption>;
  /** Represents an association between a variant and a selling plan. Selling plan allocations describe which selling plans are available for each variant, and what their impact is on pricing. */
  readonly sellingPlanAllocations: SellingPlanAllocationConnection;
  /** The Shop Pay Installments pricing information for the product variant. */
  readonly shopPayInstallmentsPricing: Maybe<ShopPayInstallmentsProductVariantPricing>;
  /** The SKU (stock keeping unit) associated with the variant. */
  readonly sku: Maybe<Scalars['String']['output']>;
  /** The in-store pickup availability of this variant by location. */
  readonly storeAvailability: StoreAvailabilityConnection;
  /** Whether tax is charged when the product variant is sold. */
  readonly taxable: Scalars['Boolean']['output'];
  /** The product variant’s title. */
  readonly title: Scalars['String']['output'];
  /** The unit price value for the variant based on the variant's measurement. */
  readonly unitPrice: Maybe<MoneyV2>;
  /** The unit price measurement for the variant. */
  readonly unitPriceMeasurement: Maybe<UnitPriceMeasurement>;
  /** The weight of the product variant in the unit system specified with `weight_unit`. */
  readonly weight: Maybe<Scalars['Float']['output']>;
  /** Unit of measurement for weight. */
  readonly weightUnit: WeightUnit;
};
// ...
export type CollectionProductsFragment = { readonly __typename: 'Collection', readonly id: string, readonly title: string, readonly products: { readonly __typename: 'ProductConnection', readonly edges: ReadonlyArray<{ readonly __typename: 'ProductEdge', readonly cursor: string, readonly node: { readonly __typename: 'Product', readonly availableForSale: boolean, readonly id: string, readonly tags: ReadonlyArray<string>, readonly title: string, readonly totalInventory: number | null, readonly featuredImage: { readonly __typename: 'Image', readonly url: any } | null, readonly selectedOrFirstAvailableVariant: { readonly __typename: 'ProductVariant', readonly id: string, readonly price: { readonly __typename: 'MoneyV2', readonly amount: any, readonly currencyCode: CurrencyCode } } | null } }>, readonly pageInfo: { readonly __typename: 'PageInfo', readonly hasNextPage: boolean, readonly endCursor: string | null } } };
// ...
export type ProductFragment = { readonly __typename: 'Product', readonly availableForSale: boolean, readonly id: string, readonly tags: ReadonlyArray<string>, readonly title: string, readonly totalInventory: number | null, readonly featuredImage: { readonly __typename: 'Image', readonly url: any } | null, readonly selectedOrFirstAvailableVariant: { readonly __typename: 'ProductVariant', readonly id: string, readonly price: { readonly __typename: 'MoneyV2', readonly amount: any, readonly currencyCode: CurrencyCode } } | null };